requirements.txt
click==8.1.7
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
selenium==4.26.1
bs4==0.0.2
beautifulsoup4==4.12.3
cryptography==43.0.3
openpyxl
textfsm
xmltodict==0.14.2
zeep==4.3.1
netmiko==4.4.0